Refactor: index.ts 추가
