+++ src/components/domain/register/index.ts
@@ -1,0 +1,2 @@
+export { default as StepCircle } from './StepCircle';
+export { default as StepProgressBar } from './StepProgressBar';
